+++ src/components/SearchBox.js
@@ -4,7 +4,7 @@ const SearchBox = ({searchChange}) => {
 	return (
 		<div className="pa2">
 			<input 
-				className="pa3 ba b--blue bg-lightest-blue"
+				className="pa3 mb2 ba b--blue bg-lightest-blue"
 				type="search" 
 				placeholder="Robot name"
 				// In here the callback function received is assigned to an event listener: 'onChange'
