+++ src/components/scroll.css
@@ -1,0 +1,4 @@
+.scroll {
+	overflow-y: scroll;
+	height: 800px;
+}
