+++ src/index.css
@@ -5,9 +5,17 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background: rgb(255,184,0);
+background: linear-gradient(90deg, rgba(255,184,0,1) 0%, rgba(255,158,0,1) 50%, rgba(255,184,0,1) 100%);
 }
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
+}
+
+.title {
+	font-family: 'Monoton';
+	font-size: 4rem;	
+	font-weight: 400;
 }
